requirements.txt
streamlit>=1.30.0
numpy>=1.21.0
tensorflow>=2.10.0
Pillow>=9.0.0
opencv-python-headless>=4.5.0
pandas>=1.3.0
scikit-learn>=1.0.0
requests>=2.26.0
groq>=0.4.0
fpdf>=1.7.2
